Update docker command for interactive bash session
This change updates the documentation to reflect that the `ENTRYPOINT` needs to be explicitly set to `bash`, otherwise the container execution will interpret "bash" as an argument.

I believe this could also be resolved via changes to the Dockerfile: changing `ENTRYPOINT` to `CMD` will allow the behavior you describe in the doc. I'm not sure of the exact use case, but [if you're trying to set the _default_ but overridable execution flow](https://docs.docker.com/engine/reference/builder/#cmd) then `CMD` might be a better fit.